requirements.txt
Pygments==2.3.1
docutils==0.14
livereload==2.6.0
pylint==2.2.2
pytest==5.2.1
pyxcute==0.5.2
twine==1.12.1
wheel==0.32.3
